AutoPlayList Rehaul
Now has a personal auto play list for each unique user. This is created
by loading the file on start and is updated while the program is
running. The data structure for storing this is a dictionary of lists,
mapping the discord id to the list of URLs that person likes. This
version currently has a lot of debug info that will soon be cleaned up.
Also added a mylist command for viewing all the songs on your personal
auto play list. ATM it prints this to the channel but will later be sent
via a PM. Use the like command to add someone else's song to your APL or
dislike to remove it from your own APL. Also updated add/remove to APL
functions.