PM the mylist, random APL selection optimizations
Instead of displaying the whole list in the text channel, sends a txt
file to the user with their personalized autoplaylist. Before, it would
randomly select a user, then randomly select a song from their list,
THEN check if the user is in the channel. Makes more sense to check if
the user in the channel before looking at their list.