Beginnings of stats feature, dislike/skip work
Nate's new stats command is added, which shows a little leaderboard of
who has the most songs in their autoplaylist! Show off who has a more
diverse selection of music. I've added a new boolean that allows you to
autoskip a song you used to be the author of, once you've disliked it
(losing the autoskip feature).