Volume improvement, patchnotes command
By popular demand, the volume will return back to default after each
song. When you're jamming to a really great song and you forgot to
change the volume back at the end, you don't want to blow your speakers
to a mediocre song. So, this is now taken care for you. Remember you can
always change your default volume in the config file! Also, we introduce
the patchnotes command. This command will allow you to see the latest
commit message (like this one) to catch up on the latest features.