OOP, not oops! Though most of it is broken
This is stage 1 of upgrading the music bot from the old string set up to
the new object oriented style. (Why didn't we do this before? because
let's be real we didn't think it would be this serious). This includes a
new User class which replaces the old dictionaries and a new Music class
which replicates a song. Instead of storing everything we know about a
song in a string and splitting with increasing, unique identifiers, we
can now add whatever we want as variables without any worry. This will
make adding features MUCH easier, it's just a matter of completing the
transition. We also introduce pickle! If you're not familiar, there is
some really good information online pertaining to python serialization.
This really comes in handy when we have large lists of complex types
such as classes. How would we have written these to a file with strings?
I dunno. But I'm glad we can pickle them and just load them whenever we
restart the bot! That's all for now, more updates to come.