Putting FUN in functionality
Adds pickling for instant start up, records the number of times a song
is played in your mylist, fixes a lot of features caused by updating the
infrastructure to OOP such as mood, tag show, tag add/remove, like,
dislike, etc. There are excessive print statements and still more to do
but "basic" functionality is back! Ghost is still broken, and there's
still a bug when only one person likes a song and they dislike it.
Remember, all issues can be found in the issues section on github.