Smarter playing
First searches the list of songs we already have added and if the search
criteria is contained by one of our song's title, then we will play that
one before searching youtube. This is useful when we the recommended
video has a long intro (we don't want that) so we want to use whatever
version we have saved. This ISN'T useful when we have generic search
critera that matches to a song that we didn't want. This feature is in
beta, it can be reverted or made to be less sensitive if it ends up
hurting more than it helps. Oh yeah and we add the title of a song while
it's being played if it had no title before.