playcount tweaks
playcount now shows plays for all users and this is done very poorly in
the backend. i havent figured out why the songs in the user's song list
aren't pointers to the songs in the APL, but that will come later. for
now, this functions as intended from user's perspective. also notify
likers prints an optional error message.